MA-22748 Software scale UI resolution fail on evk_93 board
When use below bootargs to scale down the UI resolution to make UI more smooth
on evk_93 board, the display cannot work normally because framebuffer size
checking failed in kernel DRM framework.
setenv append_bootargs androidboot.lcd_density=120 androidboot.gui_resolution=ssw640x480

Refine the operation of the new display config adding and device composition
checking to fix this issue.

Reviewed-on: http://androidsource.nxp.com/project/25592
Reviewed-on: http://androidsource.nxp.com/project/25731

+++ display/hwcomposer3/ClientFrameComposer.cpp
@@ -664,7 +664,7 @@ HWC3::Error ClientFrameComposer::setPowerMode(Display* display, PowerMode mode) 
     auto err = client->setPowerMode(displayId, power);
     if (err != HWC3::Error::None) {
         ALOGE("%s: display:%" PRIu64 " failed to set power mode:%d" PRIu64, __FUNCTION__, displayId,
-              mode);
+              static_cast<int>(mode));
     }
 
     return HWC3::Error::None;
